# biblionacho-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biblionacho API",
        "description": "Listado de URI'S para la API de Biblionacho",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/lendbooks/index": {
            "get": {
                "tags": [
                    "LendBooks"
                ],
                "summary": "Index Lend Books",
                "description": "Show all lend books",
                "operationId": "IndexLendBooks",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer id|YOUR_ACCESS_TOKEN_HERE"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ok 200"
                                        },
                                        "lendbooks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LendBook"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "errors": [
                                                    "Internal server error!"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lendbooks/store": {
            "post": {
                "tags": [
                    "LendBooks"
                ],
                "summary": "Create Lend Books",
                "description": "Create a lend book",
                "operationId": "CreateLendBook",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer id|YOUR_ACCESS_TOKEN_HERE"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Lend Book data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identification",
                                    "isbn",
                                    "observations",
                                    "returned",
                                    "user_id",
                                    "book_id"
                                ],
                                "properties": {
                                    "identification": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "yI47e2b87c"
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "returned": {
                                        "type": "boolean",
                                        "example": "false"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully registered"
                                        },
                                        "lendbook": {
                                            "$ref": "#/components/schemas/LendBook"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The isbn field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "errors": [
                                                    "Internal server error!"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LendBook": {
                "required": [
                    "identification ",
                    "isbn",
                    "observations",
                    "deadline",
                    "returned",
                    "user_id",
                    "book_id"
                ],
                "properties": {
                    "identification": {
                        "description": "identification of user",
                        "type": "string",
                        "example": "123456789"
                    },
                    "isbn": {
                        "description": "ISBN of the lend book",
                        "type": "string",
                        "example": "yI47e2b87c"
                    },
                    "observations": {
                        "description": "Observations of the lend book",
                        "type": "string",
                        "example": ""
                    },
                    "deadline": {
                        "description": "Deadline of the lend book",
                        "type": "string",
                        "format": "date",
                        "example": "10-05-2024"
                    },
                    "returned": {
                        "description": "Returned of the lend book",
                        "type": "boolean",
                        "example": false
                    },
                    "user_id": {
                        "description": "User ID of the lend book",
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "description": "Book ID of the lend book",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    }
}